ENH: Seperate tests created
