+++ TCIABrowser/Testing/Python/CMakeLists.txt
@@ -1,2 +1,2 @@
-
-#slicer_add_python_unittest(SCRIPT ${MODULE_NAME}ModuleTest.py)
+slicer_add_python_unittest(SCRIPT TCIABrowserModuleAPIV1Test.py )
+slicer_add_python_unittest(SCRIPT TCIABrowserModuleAPIV3Test.py )
